Add EmojiField
